Refactor services and remove unused utils
