--- src/resources/constants/endpoints.ts
@@ -1,2 +1,0 @@
-export const home = '/';
-export const users = '/users';
