+++ src/routes.ts
@@ -2,12 +2,15 @@ import { Router } from 'express';
 
 import acl from './middlewares/acl';
 import * as validate from './middlewares/validate';
+import authenticate from './middlewares/authenticate';
+import validateRefreshToken from './middlewares/validateRefreshToken';
+
 import * as homeController from './controllers/home';
 import * as userController from './controllers/user';
 import * as authController from './controllers/auth';
-import authenticate from './middlewares/authenticate';
+import * as passwordController from './controllers/password';
+
 import { userPOSTSchema } from './validators/userRequest';
-import validateRefreshToken from './middlewares/validateRefreshToken';
 import { loginSchema, resetSchema, forgotSchema } from './validators/authRequest';
 
 const router: Router = Router();
@@ -17,8 +20,9 @@ router.get('/', homeController.index);
 router.post('/login', validate.schema(loginSchema), authController.login);
 router.post('/refresh', validateRefreshToken, authController.refresh);
 router.post('/logout', validateRefreshToken, authController.logout);
-router.post('/forgot', validate.schema(forgotSchema), authController.forgot);
-router.post('/reset', validate.schema(resetSchema), authController.reset);
+
+router.post('/forgot', validate.schema(forgotSchema), passwordController.forgot);
+router.post('/reset', validate.schema(resetSchema), passwordController.reset);
 
 router.get('/users', authenticate, acl, userController.index);
 router.post('/users', authenticate, acl, validate.schema(userPOSTSchema), userController.store);
